backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
SQLAlchemy==2.0.21
requests==2.31.0
python-dotenv==1.0.0
geopy==2.4.0
shapely==2.0.2